Added bow and arrow meshes. And still holding objects when animation switching.

+++ Frosty/src/Frosty/RenderEngine/Renderer.hpp
@@ -60,7 +60,7 @@ namespace Frosty
 		static void UppdateEntity (const size_t& matID,ECS::CMaterial* mat, const std::string& meshName, std::shared_ptr<VertexArray> vertexArray, const size_t& transformID, ECS::CTransform* transform);
 		
 		//Suggestion: Overload this for each individual trait. That way we need not replace everything.
-		static void ChangeEntity (const int& OldMatID,ECS::CMaterial* mat, const std::string& OldMeshName, ECS::CMesh* mesh, const int& transformID, ECS::CTransform* transform);
+		static void ChangeEntity (const size_t& OldMatID,ECS::CMaterial* mat, const std::string& OldMeshName, ECS::CMesh* mesh, const size_t& transformID, ECS::CTransform* transform);
 		static void UpdateCMesh(const int& entityID, ECS::CMesh* mesh);
 
 		inline static void Shutdown() { delete s_SceneData; }
